backend-flask/requirements.txt
flask
flask-cors

opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp-proto-http 
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-requests

watchtower

aws-xray-sdk

blinker
rollbar

Flask-AWSCognito

psycopg[binary]
psycopg[pool]